chat-api/requirements.txt
chromadb
unstructured 
langchain_community 
pymupdf 
langchain-google-genai 
langchain  
pinecone 
fastapi 
uvicorn
python-dotenv
